translator: remove the bogus function mode given to CL:RANDOM
The function mode of fixnum was originally given to the Maclisp RANDOM,
which always returned a fixnum and which was used for the translation
of Macsyma's random.  Since the port to Common Lisp this fixnum mode
has been incorrectly given to CL:RANDOM and this could cause bogus
translations if (for whatever reason) CL:RANDOM was used.  For example,
translating the following and calling foo would cause a lisp error:

  foo () := ?random (1.0) / 2

See also commit 629b812f.

No problems with the test suite or share test suite.
tests/rtest_translator.mac runs as expected.