Docs: Correct (and expand) the entry for tr_bound_function_applyp
Since commit 1263968e in 2005, the docs have incorrectly stated that
tr_bound_function_applyp does not affect the generated code.  This
flag can affect on both the form and behavior of the translated code,
so I've removed the incorrect statement, expanded the description a
bit and added an example.

The docs have also failed to mention that this flag only has an effect
when tr_function_call_default is general.