Update AUTHORS file (with the skript admin/list_authors.sh)
That one gets the contributors from the git logs. Thanks to
Robert Dodier for information about the last missing names
(which I did not know).

I hope, that is better than the version from 13 years ago.
In the old file there was the text:
"As of April 2007, the Maxima developers are:".