Fix bug #3417: sign(1/zero) => 0 (where equal(zero,0))
Like the bug report title says, if we assume equal(zero,0) then we
had sign(1/zero) returning zero instead of signalling an error.

In SIGN1, COMPSPLT would determine that 1/zero is positive iff zero
is positive.  DCOMPARE then compared zero to 0, which set SIGN to
zero.  Previously at this point if SIGN was pos, neg or zero then
SIGN1 returned this sign.  Now we only compare SIGN to pos and
neg because if SIGN is zero here then more work needs to be done.

This sign(1/zero) => zero behavior was introduced in commit 6d1b6fe9
(from 2016) which changed the behavior of COMPSPLT.

No problems with the test suite or share test suite.