Remove the unused optional parameter for TRANSLATE-FILE
Under Maclisp this TTYMSGSP parameter was mandatory and the sole call
to TRANSLATE-FILE just passed the value of tr_file_tty_messagesp.  The
Common Lisp version of TRANSLATE-FILE made this parameter optional
(with the default value of tr_file_tty_messagesp), but no call ever
supplied an argument for it.

The old commented-out pre-CL code was removed in commit fe16647d.

Now just check tr_file_tty_messagesp directly in TRANSLATE-FILE since
this flag is the documented way to control this extra printing to
standard output anyway.

No functional changes.