Remove the tr_warn_* variables from the tr_state_vars list
tr_state_vars should (and is documented to) only contain the names of
translator variables which actually affect the form of the translated
code.  I sometimes find it useful to examine this list while working
on the translator, and translate_file and compile_file write a comment
in the output file that contains the names and values of the variables
in this list.

The tr_warn_* variables do not affect the translated code.  These
variables have been on this list for a long time (before Maxima 5.0),
but I'm removing them now because they are not relevant and they just
get in the way.  (The list of tr_warn_* variables wasn't even complete
anyway: tr_warn_bad_function_calls was missing from tr_state_vars.)

See also commits 576ce45c and 8e438237 which added two variables to
this list.