Remove the obsolete function TRANSLATE-ATOMS
This function was originally used to declare variables that appeared
throughout certain forms as special.  With the removal of obsolete
code in commit 231a1542 this function is now obsolete as well.

No functional changes.

No problems with the test suite or share test suite.
tests/rtest_translator.mac runs as expected.