Fix bug #3642: Lisp error when translating assume
A lisp error would occur during translation when translating an
assume call with an atomic argument (the translator had incorrectly
assumed that CAAR could be called on every argument).

It appears that this has never worked in Maxima under Common Lisp,
although it worked "accidentally" under Maclisp.  In Maclisp no error
checking was done in CAR and CDR calls in compiled code (there were
CAR and CDR variables to control the behavior in interpreted code),
so if Macsyma was compiled then calling CAAR on an atom would not
cause an error.  See also commit 6272a074 for a similar situation
in another translator fix.

No problems with the test suite or share test suite.
tests/rtest_translator.mac runs as expected, with new tests.