Slightly rewrite some rtest_translator tests to avoid FUNCALL
Several of the first rtest_translator tests use an explicit FUNCALL
to call the translated functions.  This was presumably to make sure
the functions were actually translated and that the translated
function was actually being called.

Now that translate_or_lose is used to make sure the translation
succeeds (see commit b5e69a3e), let's just use the usual Maxima
function call syntax to call the functions instead of FUNCALL.
All of the other tests use the usual syntax for calls.