Fix the bogus translation of errcatch
errcatch always yields a (Maxima) list but the mode for any errcatch
form was always the same as the mode of the last subform.  This is
not correct in general, so lisp errors could easily occur.  Consider
foo defined by:

  foo () := 1 + errcatch (2)

If foo was translated then the addition was incorrectly translated
based on both arguments being fixnums.

This is an old bug that was present in Macsyma.

The mode of an errcatch form is now any since we don't currently have
a better, more specific mode that we can use here.

No problems with the test suite or share test suite.
tests/rtest_translator.mac runs as expected, with new tests.