Replace special variable BAS with BASE in lexical extent.
This is in part a workaround for a bug in some versions of SBCL
(special variables sometimes mishandled) and in part a special
variable cleanup (special variable unnecessary in this case).
BAS is a special variable in src/mrgmac.lisp, where it is used
by some functions related to the assume database.

Fixes SF bug #3464: "Maxima compiled with SBCL-1.4.11 causes an error"