Bug fix: translating multiple functions containing local caused an error
This is a very old bug.

(%i1) foo () := local ()$ /* just something with local */
(%i2) bar () := local ()$ /* just something (else) with local */

(%i3) translate (foo);
(%o3) [foo]

(%i4) translate (bar);
error: there is already a 'local' in this block.
error: failed to translate bar
(%o4) []

Attempting to translate foo twice would also give the same error.

The problem was that the LOCAL flag was not bound by TR-LAMBDA, so it
would get set to T and never get unset.  This flag is meant to tell
the translator whether or not a local is present in the body of the
current block or function being translated.

The problem did not occur if local was within a block because LOCAL
was bound by TR-MPROG-BODY.

Now bind LOCAL in both TR-LAMBDA and BIND-TRANSL-STATE.

No problems with the test suite or share test suite.
tests/rtest_translator.mac runs as expected, with a new test.