Fix the inconsistent (wrong) return value of translated ldisp/ldisplay
The order of the elements in the list returned from ldisp and ldisplay
in the interpreted case matches the order of the args given to ldisp
and ldisplay.  I think this is the most obvious and useful order.

In the translated case, the list returned by ldisp and ldisplay has
been in the reverse order.

This is a very old bug.