Fix bug #3650: Set the datafile separator for gnuplot in draw
Set gnuplot's datafile separator to whitespace to override any
separator the user has set.

This is based on the suggestion of the bug reporter, who had the
separator set to a comma in their gnuplot configuration file.  This
caused the plots to be wrong when setting the terminal to png, jpg,
etc.

I was able to reproduce the problem by setting the separator to a
comma in my gnuplot config file, and setting the separator to
whitespace from the maxout.gnuplot file that draw writes fixes the
problem for me.