Fix up some incorrect text in quadpack docs
In a few places in the quadpack docs random bits of text were showing
up in the html docs.  This was caused by the Examples section of
commented out blocks of code, but some of the formulas were broken
over two lines and the second line didn't have a `@c` to mark the line
as a comment.  This line showed up in the output.

Just merge the line with the previous line.