Make polynomialp a built-in function
Move polynomialp from share/linearalgebra to src so it's always
available in maxima without having to load up all of linearalgebra to
get it.

We also moved the documentation from linearalgebra.texi to
Polynomial.texi.

The autoload feature is removed for polynomialp (and nonnegintegerp)
in max_ext.lisp.   And load-linearalgebra-lisp-files.lisp no longer
loads polynomialp.lisp since it's not there anymore.