Remove the ARG-LIST property when removing transl properties
The ARG-LIST property is added to a symbol's plist when defining
a function via DEFMFUN.  Some translated functions get defined with
DEFMFUN, so this should be removed when removing other translator
properties (like when killing the function).

No problems with the test suite or share test suite.
tests/rtest_translator.mac runs as expected.