Add the IMPL-NAME property when defining a function with DEFMFUN
When a function $FOO is defined using DEFMFUN, we currently have
that another internal function FOO-IMPL is also defined.  Now we
save the name of this internal function on the plist of $FOO.

This is needed for some upcoming bugfixes.

No problems with the test suite or share test suite.  I noticed
no appreciable change in the test suite run time.
tests/rtest_translator.mac runs as expected.