Work around some nonconforming behavior in GCL
Commit 8f7dc535 fixed a bug by removing some obsolete definitions
when translated functions are killed or redefined.  One of these
definitions was for a compiler macro.

Unfortunately GCL 2.6.12 doesn't know how to set a compiler macro
function with setf, so use SI:UNDEF-COMPILER-MACRO for GCL.

This was briefly discussed on the mailing list in the thread "GCL
Build seems to be broken".  I also posted about this on the gcl-devel
mailing list with the subject "(setf compiler-macro-function)", but
it's been over a week and there has been no reply.