Manually fix the output from fft routines
The output for fft(L) appears to be the conjugate of the actual output
so update the results to match the current implementation.

Not sure why, but I suspect these were created from an early version
of the fft routines which implemented the engineering definition
where the forward fft is exp(-2*%pi*%i*f).  But the original fft
version use the mathematical (?) definition of exp(+2*%pi*%i*f).  The
fft implementation was changed to match the original definitions.