DEFMTRFUN: use DEFUN to define regular functions with rest args
In commit 408b6601 the old Maclisp-style lexpr constructs were
replaced with CL-style rest args for functions and macros with rest
args that are defined with DEFMTRFUN.

DEFMTRFUN has been using DEFMFUN to define the functions with rest
args.  We do a check on the number of args before constructing the
Maxima list for the rest arg, but the change in the aforementioned
commit caused DEFMFUN to do an extra check on the number of args
and it caused the introduction of an extra function call.

Now just use DEFUN instead of DEFMFUN in the case of functions with
rest args.  DEFMFUN is still used for functions without rest args so
we still get arg checks in that case.

No problems with the test suite or share test suite.
tests/rtest_translator.mac runs as expected.  Commit 4e94033a added
tests which check the error handling for the wrong number of args in
the rest arg case.