Changelogs for 5.43 and 5.44 were missing from the latest tarball releases.
Added them to Makefile.am.

Additionally created a ChangeLog-5.45.md for the (not yet released)
next Maxima version and included that in Makefile.am, so that that does
not happen for the next release.
Removed (nearly empty) ChangeLog-Current.md.

Maybe we should remove some older "ChangeLog"-files and just keep
the 2 or three most recent? Nearly half of the files in the
toplevel directory are now called "ChangeLog-5.xx".
If one is interested, what changed in Maxima 5.10.0 (from 2006!), one
can download older tar distributions or git and change to an older
git tag to get that (old) files.