Include files like COPYING, AUTHORS in the Windows installer.
They are not installed by default, there exists an extra
make target for installing them.
Call 'make extradocinstall' when building the Windows installer.

On the mailing list, there was a discussion, that Maxima on
startup recommends reading the file 'COPYING', but it is not
installed by default (and therefore was not included in the Windows
installer).