Non-ascii filename support for plotting, operatingsystem and numericalio on sbcl+MSW
The long form: If maxima internally uses a different encoding than a
filename in the filesystem does file operations might fail on SBCL
if non-ascii chars are involved and the filename isn't converted to
its native representation. since on Win10 harmless-looking directory
names like Documents can hide a path that contains non-ascii characters
if the user's name contains them the reason for this might not be obvious
for end users.

There are many places where maxima accesses files so there would be
many additional places to fix. But I don't own a windows machine
so I hope someone else will do so and be able to test these fixes in
an easier fashion.