fmin_coblya sets ierr = -1 if constraints might not be satisfied
If the MAXCV value computed by COBYLA exceeds RHOEND, then the return
code is set to -1 to indicate that the constraints might not be
satisfied.

This is a small incompatible change with the existing function that
only said ierr could be 0, 1, or 2.  But this is a nice mostly
compatible way to indicate something bad may have happened and the
user should investigate the solution and constraints carefully.

Also cleaned up some of the comments in cobyla-interface.lisp to match
the implementation.

Update the docs to match what we do now.