Fix  bug in (bigfloat:expt a b)  when b is a bigfloat type
As an optimization for accuracy, (bigfloat:expt a b) would check to see if
the b was an integer by truncating b.  However b could be a complex
number so this would produce an error.

Fix this by checking to see if b is real or not.  If not, we use the
general expression to compute expt.

Add one test for this using zeta(3b0+%i).  This now produces a full
bigfloat value instead of an error.