Fix bogus translation of conditionals with a true test, false consequent
This is an ancient bug that was present in Macsyma.

When translating a conditional expression with a test that translated
to T and consequent that translated to NIL, that entire clause was
being skipped over.

An example of the bad behavior: the translation of

  if true then
    false
  else
    'wtf

would yield wtf instead of false.  It was not necessary for the test
and consequent to be a literal true and false, respectively; these
just needed to translate to T and NIL.

Since this was around in the old Macsyma days, my guess is that this
was a careless attempt to save some conses.  The idea being that we
can omit the final clause in

  (COND ... (T NIL))

The problem is that the code to handle this optimization was looking
at all of the clauses instead of just the final one.

We'll keep this tiny optimization for now, but we'll restrict this to
the final clause.

No problems with the test suite or share test suite.
tests/rtest_translator.mac runs as expected, with new tests.