Bug fix: gcfactor(x*%i) => lisp error
The commit message for commit d7ff7c39 claims it was only inlining
a function, but that commit made several other changes as well.
One of the changes was using = instead of EQUAL in GCDISP.

The use of = here causes various things like gcfactor(x*%i) and
gcfactor(%i/2) to signal lisp errors.

I'm changing this = to EQL in GCDISP.  (EQUAL isn't necessary here;
the original use of EQUAL goes back to the Maclisp days when EQL
didn't exist.)

No problems with the test suite or share test suite.  New tests are
added to rtest_numth.