(Partly) revert commit aceb050.
This commit mentioned only the Windows installer related changes,
and is superseded with the Changelog Gunter created with committed [06f0cb].

Kept the missing changelog entries in 'make dist'. And used my description
for the Windows related changes, for an ordinary user the NSIS syntax
'RequestExecutionLevel admin' is not really important.