Fix bug #3704: Translator gives internal error
An error could occur in FREE-LISP-VARS due to a backquoted form from
the translation of MFACTORIAL.  Whether or not an error occurred was
implementation dependent because it depended on how the backquote
reader macro is implemented.  I observed the error with SBCL, but not
with ECL or CLISP.

So eliminate that backquoted form from the translation of MFACTORIAL
so the naive FREE-LISP-VARS doesn't choke on it.

No problems with the test suite or share test suite.
tests/rtest_translator.mac runs as expected, with new tests.