doc/info/category-macros.texi: correct definition of mmref
In commits e1527087 and 57819f2f, Jinsong Zhao's patch to
texinfo.tex's \ref macro was implemented.

Wolfgang Dautermann noticed the pdf manual had "funny" characters
appearing in place of _ (and others).

This was caused by Zhao's patch--it was not properly isolating its
catcode changing.

Rather than maintain a separate macro for \ref, I have patched \ref in
texinfo.tex (and asked the patch be adopted upstream) so that it can
be customized to do what the patch in the above-mentioned commits
should have done.

Some additional changes:

-\mmref was defined as a 2-arg macro, but the second argument was
never used, so it is dropped;

-\mmref needs to be defined before its dependents, so the tex block at
the end of the file is moved above the definition of @mref;

-the current patch removes the need for @mrefdot & co.; they have been
left in, but redefined to use the correct parent (@mref) rather than
duplicating code.