doc/info/texinfo.tex: update to include recent patches from texinfo
This is the version from commit 7a920ade9 from the texinfo repository.
It includes a patch that is functionally equivalent to that recently
introduced here. This overwrites the patch in fea7e4f87.

category-macros.texi has a patched version of \mmref that is
functionally equivalent to its predecessor. This amends 942fc018e.

Ref:

https://lists.gnu.org/archive/html/bug-texinfo/2021-01/msg00011.html