Fix bug #3543: bug with polynomialp
This fixes polynomialp so it compares variables with ALIKE1 instead
of EQUAL.  Barton Willis proposed this fix in the bug report.

This was also briefly discussed in the mailing list thread
"Extracting coefficients and powers from a polynomial" on
2020-09-16.

No problems with the test suite or share test suite.  A new test
has been added to rtest_polynomialp.