translator: fix the finding of free lisp vars in LET forms
There was a bug here where only free vars that were subforms of other
body forms were found.

Example: X was found to be free in

  (LET () (FOO X))

but not in

  (LET () X)

The problem is that the code for skipping go tags was copied from
the free lisp vars stuff for PROG (see commit 95ded719).

See also commit ad7ae940 which fixed a separate issue in the finding
of free lisp vars in LET forms.

While here, add and reword some comments so they're consistent with
the comments for the newer free lisp vars stuff for PROG.

No problems with the test suite or share test suite.
rtest_translator runs as expected.