translator: allow integer go tags without giving warnings
Integer go tags have been allowed but their use would cause translator
warnings (and add an unnecessary extra call to the translator just to
trivially translate the integer).

In Common Lisp go tags are allowed to be symbols or integers and they
are compared with EQL, while in Maclisp go tags are allowed to be any
atoms and they are compared with EQ.  The translator has been assuming
that go tags could only be symbols here, which was pretty reasonable
under Maclisp but incorrect under Common Lisp.

No problems with the test suite or share test suite.
rtest_translator runs as expected, with a new test.