translator: precisely check for go tags when finding free lisp vars
In Common Lisp go tags are allowed to be symbols or integers and they
are compared with EQL, while in Maclisp go tags are allowed to be any
atoms and they are compared with EQ.  The translator has been assuming
that go tags could be any sort of atom here, which was correct under
Maclisp but not strictly correct under Common Lisp.

Now use GO-TAG-P (introduced in commit bdc958b8) instead of ATOM when
checking for go tags when finding free lisp vars.

No functional changes.  It's just clearer to use GO-TAG-P instead of
ATOM.

No problems with the test suite or share test suite.
rtest_translator runs as expected.