Remove series for erf and friends
The definitions for the power series for erf and friends breaks the
taylor series for erf expanded at a non-zero point.  I don't know why,
but let's remove this and the corresponding tests until this is fixed.

Don't know why things work for the trig functions but not erf.
Perhaps taylor has info on how to do these.  Perhaps the exp-form
property on these symbols tell taylor how to do it.