Allow debugmode to be locally bound
Binding debugmode has been broken because *MDEBUG* was being bound
to NIL by MBIND via the WITH-$ERROR macro.

Instead of WITH-$ERROR binding *MDEBUG* and other variables in an
effort to force MERROR to signal a MAXIMA-$ERROR condition, let's
just have it bind the new *MERROR-SIGNALS-$ERROR-P* and have MERROR
check for this.

While here, also move the modified WITH-$ERROR so the code for the
MAXIMA-$ERROR stuff is together.

WITH-$ERROR was introduced in commit b50ab644 and has been used in
MBIND since commit 88d94f51 (both commits were in December 2004).

This was briefly discussed in the mailing list thread "debugmode:lisp"
on 2021-03-15.

No problems with the test suite or share test suite.