Fix bug #3760: apropos("") errors
This bug was caused by the existence of the empty symbol || in the
:maxima symbol table and the fact that get-first-char does not check
if the string is null. Since get-first-char seems optimized for speed,
it has been left as is, and a special case for the empty symbol has
been added in the filter.

Since the existing function, FILTER, was a single-use
re-implementation of DELETE-IF-NOT, it has been replaced with
DELETE-IF-NOT. The nested calls ($LISTIFY ($SETIFY ...)) have been
replaced with a single call to (REMOVE-DUPLICATES ...).

Finally, the documentation appeared to suggest that APROPOS could
accept either symbols or strings, while the code accepted only
strings. APROPOS now accepts either symbols or strings.

The documentation has been updated with examples and a test has been
added.