Extend laplace(t^z,t,s) to complex z
For complex z, the real part of z needs to be greater than -1.  The
only change necessary to make this work is to update LAPEXPT so it
checks the sign of the real part of z+1 instead of the sign of the
entire expression z+1.

No problems with the test suite or share test suite.  Some new tests
have been added to rtest_laplace.