limit of diff
Get rid of code to "hide" integrals and derivatives inside limit.
This led to bugs because the hidden expression was treated as
a constant.

New code to substitute the limit value into a %derivative.
The %derivative may come from a quoted 'diff or from some
diff that maxima cannot evaluate, such as diff(exp(x), x, n).

Fixes #3767 incorrect limit (radius of convergence)