imaxima.lisp: refactor wx-gnuplot-installation-check
The old code for wx-gnuplot-installed-p was wrong for gcl because it
did not use $gnuplot_command, and the gcl vs non-gcl code checked two
different things. I have retained the gist of the non-gcl code and
made it work in gcl, too, by using ignore-errors instead of
handler-bind.

Since wx-gnuplot-installed-p is only used in
wx-gnuplot-installation-check, I have put it inside the body of the
latter.

The comment about style-warning-suppressor is irrelevant because it
was removed some time ago, so the comment has been removed.