Windows installer: Compile TCL/TK statically.
For Windows/32bit I got a strange error message
when TCL wanted to load the library tcl86.dll:
"The procedure entry point AccessCheck could not
be located in the dynamic link library
C:\maxima-5.45.0\bin\tcl86.dll".

It works, when using the configure parameter
--enable-shared=no

wish86.exe is called wish86s.exe in that case.

This change is also included in the 32 bit Windows-installer
for Maxima 5.45.0.