Fix bug in function dimensions (package ezunits) in the presence of ev(..., nouns).
Disable *NOUNSFLAG* (which is enabled by ev(..., nouns)) when
processing dimensions(...), which is implemented via simplification rules
on nounified expressions; in the presence of ev(..., nouns), that scheme fails,
with stack overflow errors.

ev(..., nouns=false) is not recognized, so block([?\*nounsflag\*: false], ...) is applied instead.

*NOUNSFLAG* is enabled by the functions constructed by COERCE-FLOAT-FUN,
as part of an effort to arrive at a numerical result; the bug was first
noticed in a call to plot2d, which calls COERCE-FLOAT-FUN.

Fixes SF bug #3789: "package ezunits: ev(dimensions(u), nouns) stack overflow"