maxima.bat: Don't use variables @SBCL_NAME@ / @CLISP_NAME@.
The idea was good, but when crosscompiling the Windows installer
on Linux (which is the usual way, Windows installers are created
nowadays), the 'compiler' is a shell script (using 'wine' and the
Windows executables), so if that 'executable' is used on Windows,
it does not exist.
This also affected the nightly build.

Thanks to Paul Verheyen for reporting the issue on the Maxima mailing
list.

This reverts some code from commit 1a34052.