In limit, try harder to detect failed Taylor series expansion
In MRV-LEADTERM, look for presence of variable of interest in would-be
leading coefficient; if found there, punt to TAY-ERROR which throws
TAYLOR-CATCH when SILENT-TAYLOR-FLAG is enabled (the usual circumstance).
The effect of this is to cause limit to give up and return a noun expression.

Fixes SF bug #3826: "limit returns temp variable expression"