Cut out some fixnum declarations in rat3c.
Cut out fixnum declarations for variable K in functions PGCD2 and PGCDU1 in src/rat3c.lisp.
This avoids an error when working with some polynomials which
have coefficients which are rationals with large numerator
or denominator (derived from floating point numbers).

The test case for this bug fix is pretty large, comprising 20
integrals, all of which need to be evaluated in order for the
last one to fail; this probably shows that the error has something
to do with special variables. Also, I observed the error only
with SBCL.

Fixes SF bug #3810: "integrate error \"not of type FIXNUM\" for integrand with floats in it"