In package eigen, ensure local list is referenced in subscripted expression
Subvert global array superseding local list in subscripted expression
by writing, e.g., (+foo)[i] instead of foo[i], where foo is a global array;
the (+foo) prevents Maxima from detecting the global array.

Fixes SF bug #3654: "uniteigenvectors fails if uv[1] is used in the user main program."
This is a suboptimal bug fix. A better way to go about it is to
implement lexical symbols, which is a work in progress.