Fix printing for engineering format.
Due to round-off problems the computation of n = floor(log10(x)) where
x is the number to be printed is sometimes too small.

To solve this, we test if 10.0^n <= x < 10.0^n.  If it fails, then the
exponent is too small by 1, so we increase n by 1.  This takes care of
most cases.

However, that's not enough.  Some lisps like ecl (20.4.24) end up
still printing a number that doesn't have the right exponent:

> (format t "~,v,,ve" 15 2 1d10)
100.00000000000000d+8

So we check that the printed exponent is a multiple of 3.  If not, we
decrement the scale factor by 1 which seems to fix things.  For ecl,
we get:

> (format t "~,v,,ve" 15 1 1d10)
10.000000000000000d+9


However, ccl64 1.12 on Linux, at least, is broken and we can't do
anything about it:

? (format t "~,v,,ve" 15 2 1d10)
10.00000000000000D+8

The printed result is off by a factor of 10; the exponent should be
9.  Other lisps I tried (clisp, cmucl, gcl) print this correctly.

There may be other cases where things are wrong.  I only used

makelist(float(10^i),i,-10,10);

for testing.