MARRAYREF: fix an incorrectly constructed expression passed to ARRFIND
This is an ancient bug that was present in Macsyma.

When use_fast_arrays:false, MARRAYREF would pass an expression to
ARRFIND without the ARRAY flag in some cases.  The header would
actually contain the symbol holding the array instead of the intended
ARRAY flag.

This bug affected various things, such as arrayapply, translated array
references and translated calls to subvar.

(%i1) array (a, complete, 5)$

(%i2) arrayapply (a, [3]);
(%o2) a(3)

No problems with the test suite or share test suite.  New tests have
been added to rtest2 and rtest_translator.