MREAD-SYNERR: check for an EOF value when traversing *PARSE-WINDOW*
It makes sense to stop accumulating characters for an error message
when we find the EOF value *PARSE-STREAM-EOF* in *PARSE-WINDOW*.
This EOF value can show up in *PARSE-WINDOW* when loading a Maxima
file, for example.

I stumbled across this when I got a lisp error out of MREAD-SYNERR.
The EOF value (which is -1) wasn't checked for and this led to a
bogus CHAR= comparison.

No problems with the test suite or share test suite.