MREAD-SYNERR: don't give the tabulate format directive a negative column
This would happen when there were less than two applicable characters
in *PARSE-WINDOW* to print.

It was implementation-dependent whether or not this would cause a lisp
error.  Here was an easy way to get a lisp error from clisp:

(%i1) )
incorrect syntax: Illegal use of delimiter )
)

*** - SYSTEM::FORMAT-TABULATE: argument -1 is not a nonnegative fixnum

No problems with the test suite or share test suite.