Fix bug #2876: Error simplifying infix operators declared l/rassociative
This is an ancient bug that was present in Macsyma.

Infix operators are assumed to be binary.  In SIMPARGS there has been
a two-arg check for expressions with an infix operator, including
operators that have been declared lassociative or rassociative:

(%i1) infix ("++")$

(%i2) declare ("++", lassociative)$

(%i3) a ++ b ++ c;

++: expected exactly 2 arguments but got 3: [a, b, c]
 -- an error. To debug this try: debugmode(true);

Now the two-arg check is omitted for infix operators with a declared
associativity.  Applying such an operator to less than two args or
more than two args is no longer an error.

No problems with the test suite or share test suite.  New tests have
been added to rtest11.